Implemented changes for situation found while working on JellyBelly
See: aa78a9386843005c3ca632107dd560132fa801ff

I removed an absurd loading of partially loaded sequences. When a sequence could
not entirely be copied into the sequence buffer. As much sequence as possible
was copied and the rest of the sequence was placed once the buffer was encoded.
This made no sense as the sequence was technically already loaded in the
"memory" memebr of the sqzfastx_t object